catalog structure change
