+++ src/Sylius/Bundle/ReportBundle/Form/Type/DataFetcher/SalesTotalType.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\ReportBundle\Form\Type;
+namespace Sylius\Bundle\ReportBundle\Form\Type\DataFetcher;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
