+++ src/Sylius/Bundle/ReportBundle/Form/Type/Renderer/ChartConfigurationType.php
@@ -9,17 +9,18 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\ReportBundle\Form\Type;
+namespace Sylius\Bundle\ReportBundle\Form\Type\Renderer;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Sylius\Bundle\ReportBundle\Renderer\ChartRenderer;
 
 /**
  * Renderer configuration form type
  *
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
  */
-class TableConfigurationType extends AbstractType
+class ChartConfigurationType extends AbstractType
 {
     /**
      * {@inheritdoc}
@@ -27,10 +28,14 @@ class TableConfigurationType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
+            ->add('type', 'choice', array(
+                'label' => 'sylius.form.report.chart.type',
+                'choices' => ChartRenderer::getChartTypes(),
+            ))
             ->add('template', 'choice', array(
                 'label' => 'sylius.form.report.renderer.template',
                 'choices' => array(
-                    'SyliusReportBundle:Table:default.html.twig' => 'Default',
+                    'SyliusReportBundle:Chart:default.html.twig' => 'Default',
                 ),
             ))
         ;
@@ -41,6 +46,6 @@ class TableConfigurationType extends AbstractType
      */
     public function getName()
     {
-        return 'sylius_renderer_table';
+        return 'sylius_renderer_chart';
     }
 }
