+++ src/Sylius/Bundle/ReportBundle/spec/Form/Type/Renderer/ChartConfigurationTypeSpec.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\ReportBundle\Form\Type;
+namespace spec\Sylius\Bundle\ReportBundle\Form\Type\Renderer;
 
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\Form\FormBuilder;
@@ -18,11 +18,11 @@ use Prophecy\Argument;
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
  */
-class TableConfigurationTypeSpec extends ObjectBehavior
+class ChartConfigurationTypeSpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\ReportBundle\Form\Type\TableConfigurationType');
+        $this->shouldHaveType('Sylius\Bundle\ReportBundle\Form\Type\Renderer\ChartConfigurationType');
     }
 
     public function it_should_be_abstract_type_object()
@@ -32,11 +32,12 @@ class TableConfigurationTypeSpec extends ObjectBehavior
 
     public function it_has_name()
     {
-        $this->getName()->shouldReturn('sylius_renderer_table');
+        $this->getName()->shouldReturn('sylius_renderer_chart');
     }
 
-    public function it_builds_form_with_template_choice(FormBuilder $builder)
+    public function it_builds_form_with_type_choice_and_template_choice(FormBuilder $builder)
     {
+        $builder->add('type', 'choice', Argument::any())->willReturn($builder);
         $builder->add('template', 'choice', Argument::any())->willReturn($builder);
 
         $this->buildForm($builder, array());
